+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/residual/gluon.py
@@ -18,14 +18,26 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-"""Registration shims for AMD Gluon AttnRes kernels."""
+"""Registration shims for AMD Gluon residual kernels."""
 
 from __future__ import annotations
 
 import torch
-from tokenspeed_kernel.platform import ArchVersion, CapabilityRequirement
+from tokenspeed_kernel.ops.residual.triton import (
+    _mhc_pre_impl,
+    _mhc_prenorm_gemm_triton,
+)
+from tokenspeed_kernel.platform import (
+    ArchVersion,
+    CapabilityRequirement,
+    current_platform,
+)
 from tokenspeed_kernel.registry import Priority, register_kernel
-from tokenspeed_kernel.signature import format_signatures
+from tokenspeed_kernel.signature import (
+    dense_tensor_format,
+    format_signature,
+    format_signatures,
+)
 
 try:
     from tokenspeed_kernel_amd.ops.gfx950.attention.kda.attn_res import (
@@ -45,8 +57,8 @@ else:
 if _attn_res_rmsnorm_gfx950_impl is not None:
 
     @register_kernel(
-        "attn_res",
-        "fwd",
+        "residual",
+        "attn_res_fwd",
         name="gluon_attn_res_fwd_gfx950",
         solution="gluon",
         capability=CapabilityRequirement(
@@ -102,8 +114,8 @@ if _attn_res_rmsnorm_gfx950_impl is not None:
         )
 
     @register_kernel(
-        "attn_res",
-        "fwd",
+        "residual",
+        "attn_res_fwd",
         name="gluon_attn_res_fwd_gfx1250",
         solution="gluon",
         capability=CapabilityRequirement(
@@ -171,4 +183,58 @@ else:
         ) from _IMPORT_ERROR
 
 
-__all__ = ["gluon_attn_res_fwd_gfx950", "gluon_attn_res_fwd_gfx1250"]
+if current_platform().is_amd:
+    from tokenspeed_kernel_amd.ops.gfx950.mhc import (
+        gluon_mhc_pre_reduce_apply_gfx950 as _mhc_pre_reduce_apply_impl,
+    )
+
+    @register_kernel(
+        "residual",
+        "mhc_pre",
+        name="gluon_mhc_pre_gfx950",
+        solution="gluon",
+        capability=CapabilityRequirement(
+            min_arch_version=ArchVersion(9, 5),
+            max_arch_version=ArchVersion(9, 5),
+            vendors=frozenset({"amd"}),
+        ),
+        signatures=frozenset(
+            {
+                format_signature(
+                    residual=dense_tensor_format(torch.bfloat16),
+                    fn=dense_tensor_format(torch.float32),
+                    hc_scale=dense_tensor_format(torch.float32),
+                    hc_base=dense_tensor_format(torch.float32),
+                )
+            }
+        ),
+        traits={
+            "num_tokens": frozenset(range(1, 65)),
+            "hc_mult": frozenset({4}),
+            "hidden_size": frozenset({4096, 7168}),
+            "sinkhorn_iters": frozenset({20}),
+        },
+        priority=Priority.SPECIALIZED,
+        tags={"amd", "gfx950", "latency"},
+    )
+    def gluon_mhc_pre_gfx950(
+        residual: torch.Tensor,
+        fn: torch.Tensor,
+        hc_scale: torch.Tensor,
+        hc_base: torch.Tensor,
+        rms_eps: float,
+        hc_eps: float,
+        sinkhorn_iters: int,
+    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        """Run split-K prenorm with a GFX950 Gluon reduction/Sinkhorn stage."""
+        return _mhc_pre_impl(
+            residual,
+            fn,
+            hc_scale,
+            hc_base,
+            rms_eps,
+            hc_eps,
+            sinkhorn_iters,
+            _mhc_prenorm_gemm_triton,
+            pre_reduce_apply_impl=_mhc_pre_reduce_apply_impl,
+        )
