+++ tokenspeed-kernel/python/tokenspeed_kernel/registry.py
@@ -480,11 +480,11 @@ def load_builtin_kernels() -> None:
                 "tokenspeed_kernel.numerics.reference."
             ):
                 del sys.modules[key]
-    import tokenspeed_kernel.ops.attn_res  # noqa: F401
     import tokenspeed_kernel.ops.embedding  # noqa: F401
     import tokenspeed_kernel.ops.gemm  # noqa: F401
     import tokenspeed_kernel.ops.moe  # noqa: F401
     import tokenspeed_kernel.ops.quantization  # noqa: F401
+    import tokenspeed_kernel.ops.residual  # noqa: F401
     import tokenspeed_kernel.ops.sampling  # noqa: F401
     import tokenspeed_kernel.ops.transform  # noqa: F401
